BUGFIX: Keep default action non cacheable
Otherwise user input is saved and re displayed to foreign visitors.

If someone is in need for an cached version, he can create a plugin
inside his sitepackage with a different plugin signature but same
action.

If that does not work, we might add another action for caching.

Also keep old action name to be non breaking.